# target/classes/swagger.yaml
swagger: '2.0'
info:
  description: Documentação da API de acesso aos endpoints com Swagger
  version: '1.0'
  title: Swagger API
host: 'localhost:8080'
basePath: /
tags:
  - name: usuario-api-controller
    description: the usuario API
paths:
  /usuario:
    get:
      tags:
        - usuario
      summary: Buscar todos Usuarios
      description: endpoint que busca todos usuarios da base de dados.
      operationId: buscarTodosUsuarios
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Sucesso ao buscar todos usuarios
          schema:
            type: array
            items:
              $ref: '#/definitions/Usuario'
        '204':
          description: Nenhum Usuario encontrado na base de dados
          schema:
            type: array
            items:
              $ref: '#/definitions/Usuario'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '422':
          description: Request nao processada
    post:
      tags:
        - usuario
      summary: Adicionar um novo Usuario
      description: endpoint que cria um usuario na base de dados.
      operationId: criarUsuario
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: usuario
          description: Novo Usuario
          required: true
          schema:
            $ref: '#/definitions/Usuario'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Retorno'
        '201':
          description: Usuario inserido com sucesso
          schema:
            $ref: '#/definitions/Retorno'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '422':
          description: Request nao processada
  /usuario/logar:
    get:
      tags:
        - usuario
      summary: Logar do Usuario
      description: endpoint que loga um usuario na aplicacao.
      operationId: logarUsuario
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: usuario
          in: header
          description: Email do Usuario
          required: true
          type: string
        - name: senha
          in: header
          description: Senha do Usuario
          required: true
          type: string
      responses:
        '200':
          description: sucesso ao realizar login
          schema:
            $ref: '#/definitions/Retorno'
        '400':
          description: Campos invalidos enviados na request
          schema:
            $ref: '#/definitions/Retorno'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Usuario Nao Encontrado
          schema:
            $ref: '#/definitions/Retorno'
        '405':
          description: Erro de Validacao
          schema:
            $ref: '#/definitions/Retorno'
  /usuario/recuperar-senha:
    post:
      tags:
        - usuario
      summary: Recuperar Senha do Usuario
      description: endpoint que envia senha para o email do usuario.
      operationId: recuperarSenha
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: usuario
          in: header
          description: email do usuario a ser recuperado a senha
          required: true
          type: string
      responses:
        '200':
          description: sucesso ao reenviar email com a senha ao cliente
          schema:
            $ref: '#/definitions/Retorno'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '422':
          description: Campos invalidos enviados na request
          schema:
            $ref: '#/definitions/Retorno'
  '/usuario/{idUsuario}':
    get:
      tags:
        - usuario
      summary: 'Buscar um Usuario {ID}'
      description: endpoint que busca um usuario pelo id na base de dados.
      operationId: buscarUsuarioPorId
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: idUsuario
          in: path
          description: Id do Usuario
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Sucesso ao buscar usuario
          schema:
            $ref: '#/definitions/Usuario'
        '400':
          description: ID do Usuario invalido
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Usuario Nao Encontrado
        '422':
          description: Request nao processada
    put:
      tags:
        - usuario
      summary: Atualizar um Usuario existente
      description: endpoint que atualiza um usuario na base de dados.
      operationId: atualizarUsuario
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: usuario
          description: Usuario atualizado
          required: true
          schema:
            $ref: '#/definitions/Usuario'
        - name: idUsuario
          in: path
          description: ID do Usuario
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Atualizacao realizada com sucesso
          schema:
            $ref: '#/definitions/Retorno'
        '201':
          description: Created
        '400':
          description: ID do Usuario invalido
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Usuario Nao Encontrado
        '422':
          description: Request nao processada
    delete:
      tags:
        - usuario
      summary: Deletar um Usuario
      description: endpoint que deleta um usuario pelo id na base de dados.
      operationId: deletarUsuario
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: idUsuario
          in: path
          description: Id do Usuario
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
        '204':
          description: exclusao realizada com sucesso
        '400':
          description: ID do Usuario invalido
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Usuario Nao Encontrado
        '405':
          description: Erro de Validacao
definitions:
  Retorno:
    type: object
    properties:
      mensagem:
        type: string
      sucesso:
        type: boolean
        example: false
  RetornoTodosUsuarios:
    type: object
    properties:
      mensagem:
        type: string
      sucesso:
        type: boolean
        example: false
      usuarios:
        type: array
        items:
          $ref: '#/definitions/Usuario'
  RetornoUsuario:
    type: object
    properties:
      mensagem:
        type: string
      sucesso:
        type: boolean
        example: false
      usuario:
        $ref: '#/definitions/Usuario'
  Usuario:
    type: object
    properties:
      celular:
        type: string
      cpf:
        type: string
        description: Status Usuario
      email:
        type: string
      endereco:
        type: string
      id:
        type: integer
        format: int32
      nomeCompleto:
        type: string
      senha:
        type: string
      statusUsuario:
        type: string
        enum:
          - PENDENTE_APROVACAO
          - ATIVO
          - INATVO
          - NEGADO